Set map as home activity
