+++ app/src/main/java/com/iic/shopingo/ui/HomeActivity.java
@@ -1,6 +1,7 @@
 package com.iic.shopingo.ui;
 
 import android.content.Intent;
+import android.location.Location;
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 import android.util.Log;
@@ -12,27 +13,33 @@ import bolts.Continuation;
 import bolts.Task;
 import butterknife.ButterKnife;
 import butterknife.InjectView;
+import butterknife.OnClick;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.MapFragment;
+import com.google.android.gms.maps.model.Marker;
 import com.iic.shopingo.R;
 import com.iic.shopingo.api.ApiResult;
+import com.iic.shopingo.api.request.GetNearbyShoppersCommand;
+import com.iic.shopingo.api.request.ShoppersApiResult;
 import com.iic.shopingo.api.trip.StartTripCommand;
 import com.iic.shopingo.api.user.CurrentStateApiResult;
 import com.iic.shopingo.api.user.GetCurrentStateCommand;
 import com.iic.shopingo.dal.models.IncomingRequest;
 import com.iic.shopingo.services.CurrentUser;
 import com.iic.shopingo.services.FacebookConnector;
-import com.iic.shopingo.ui.home.ActionCardView;
+import com.iic.shopingo.services.location.CurrentLocationProvider;
+import com.iic.shopingo.services.location.LocationUpdatesListenerAdapter;
+import com.iic.shopingo.ui.home.MapManager;
+import com.iic.shopingo.ui.request_flow.activities.CreateShoppingListActivity;
 import com.iic.shopingo.ui.request_flow.activities.RequestStateActivity;
-import com.iic.shopingo.ui.request_flow.activities.SelectShopperActivity;
 import com.iic.shopingo.ui.trip_flow.activities.ManageTripActivity;
 import java.util.ArrayList;
 
-public class HomeActivity extends ActionBarActivity implements ActionCardView.Listener {
+public class HomeActivity extends ActionBarActivity implements GoogleMap.OnMarkerClickListener {
 
-  @InjectView(R.id.home_action_card_shopping)
-  ActionCardView startShoppingCard;
+  private final long REQUEST_INTERVAL = 10 * 1000; // 10 seconds in milliseconds
 
-  @InjectView(R.id.home_action_card_make_request)
-  ActionCardView makeRequestCard;
+  private MapManager mapManager;
 
   @InjectView(R.id.home_content_container)
   View contentContainer;
@@ -95,8 +102,19 @@ public class HomeActivity extends ActionBarActivity implements ActionCardView.Li
   private void init() {
     setContentView(R.layout.activity_home);
     ButterKnife.inject(this);
-    startShoppingCard.setListener(this);
-    makeRequestCard.setListener(this);
+
+    GoogleMap map =  ((MapFragment) getFragmentManager().findFragmentById(R.id.select_shopper_map)).getMap();
+
+    map.setOnMarkerClickListener(this);
+
+    new CurrentLocationProvider(this, REQUEST_INTERVAL, new LocationUpdatesListenerAdapter() {
+      @Override
+      public void onLocationUpdated(Location location) {
+        mapManager.setUserLocation(location);
+      }
+    });
+
+    mapManager = new MapManager(this, map);
   }
 
   @Override
@@ -124,6 +142,42 @@ public class HomeActivity extends ActionBarActivity implements ActionCardView.Li
     return super.onOptionsItemSelected(item);
   }
 
+  @OnClick(R.id.select_shopper_go_yourself_button)
+  public void onGoYourself(View view) {
+    ApiTask<ApiResult> task = new ApiTask<>(getSupportFragmentManager(), "Starting trip...", new StartTripCommand(CurrentUser.getToken()));
+
+    task.execute().continueWith(new Continuation<ApiResult, Void>() {
+      @Override
+      public Void then(Task<ApiResult> task) throws Exception {
+        if (!task.isFaulted() && !task.isCancelled()) {
+          CurrentUser.getInstance().state = CurrentUser.State.TRIPPING;
+          Intent intent = new Intent(HomeActivity.this, ManageTripActivity.class);
+          intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TASK);
+          startActivity(intent);
+        } else {
+          Toast.makeText(HomeActivity.this, "Could not start trip: " + task.getError().getMessage(),
+              Toast.LENGTH_LONG).show();
+        }
+        return null;
+      }
+    }, Task.UI_THREAD_EXECUTOR);
+  }
+
+  private void updateShoppers() {
+    GetNearbyShoppersCommand req = new GetNearbyShoppersCommand(CurrentUser.getToken());
+    req.executeAsync().continueWith(new Continuation<ShoppersApiResult, Object>() {
+      @Override
+      public Object then(Task<ShoppersApiResult> task) throws Exception {
+        if (!task.isFaulted() && !task.isCancelled()) {
+          mapManager.setShoppers(task.getResult().shoppers);
+        } else {
+          // TODO: handle failure
+        }
+        return null;
+      }
+    }, Task.UI_THREAD_EXECUTOR);
+  }
+
   private void navigateToOnboarding() {
     Intent intent = new Intent(this, OnboardingActivity.class);
     intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
@@ -139,30 +193,14 @@ public class HomeActivity extends ActionBarActivity implements ActionCardView.Li
   private void showContent() {
     spinnerContainer.setVisibility(View.GONE);
     contentContainer.setVisibility(View.VISIBLE);
+    updateShoppers();
   }
 
   @Override
-  public void onCardClicked(ActionCardView cardView) {
-    if (cardView == startShoppingCard) {
-      ApiTask<ApiResult> task = new ApiTask<>(getSupportFragmentManager(), "Starting trip...", new StartTripCommand(CurrentUser.getToken()));
-
-      task.execute().continueWith(new Continuation<ApiResult, Void>() {
-        @Override
-        public Void then(Task<ApiResult> task) throws Exception {
-          if (!task.isFaulted() && !task.isCancelled()) {
-            CurrentUser.getInstance().state = CurrentUser.State.TRIPPING;
-            Intent intent = new Intent(HomeActivity.this, ManageTripActivity.class);
-            startActivity(intent);
-          } else {
-            Toast.makeText(HomeActivity.this, "Could not start trip: " + task.getError().getMessage(),
-                Toast.LENGTH_LONG).show();
-          }
-          return null;
-        }
-      }, Task.UI_THREAD_EXECUTOR);
-    } else {
-      Intent intent = new Intent(this, SelectShopperActivity.class);
-      startActivity(intent);
-    }
+  public boolean onMarkerClick(Marker marker) {
+    Intent intent = new Intent(this, CreateShoppingListActivity.class);
+    intent.putExtra(CreateShoppingListActivity.EXTRAS_SHOPPER_KEY, mapManager.contactFromMarker(marker));
+    startActivity(intent);
+    return true;
   }
 }
