+++ app/src/main/java/com/iic/shopingo/ui/request_flow/activities/RequestStateActivity.java
@@ -67,7 +67,7 @@ public class RequestStateActivity extends ActionBarActivity {
           request.setStatus(BaseRequest.RequestStatus.CANCELED);
           CurrentUser.getInstance().state = CurrentUser.State.IDLE;
           CurrentUser.getInstance().save();
-          finishAndStartActivity(SelectShopperActivity.class);
+          finishAndStartActivity(HomeActivity.class);
         } else {
           Toast.makeText(RequestStateActivity.this, "Could not cancel request: " + task.getError().getMessage(), Toast.LENGTH_LONG).show();
         }
@@ -100,7 +100,7 @@ public class RequestStateActivity extends ActionBarActivity {
   @Optional
   @OnClick(R.id.request_state_try_again_button)
   public void onTryAgain(View view) {
-    finishAndStartActivity(SelectShopperActivity.class);
+    finishAndStartActivity(HomeActivity.class);
   }
 
   @Optional
