--- app/src/main/java/com/iic/shopingo/ui/request_flow/views/SelectShopperListItemView.java
@@ -1,70 +1,0 @@
-package com.iic.shopingo.ui.request_flow.views;
-
-import android.content.Context;
-import android.location.Location;
-import android.util.AttributeSet;
-import android.view.LayoutInflater;
-import android.view.ViewGroup;
-import android.widget.FrameLayout;
-import android.widget.ImageView;
-import android.widget.TextView;
-import butterknife.ButterKnife;
-import butterknife.InjectView;
-import com.iic.shopingo.R;
-import com.iic.shopingo.dal.models.Contact;
-import com.squareup.picasso.Picasso;
-
-public class SelectShopperListItemView extends FrameLayout {
-  @InjectView(R.id.select_shopper_list_item_photo)
-  ImageView photoView;
-
-  @InjectView(R.id.select_shopper_list_item_name)
-  TextView nameView;
-
-  @InjectView(R.id.select_shopper_list_item_distance)
-  TextView distanceView;
-
-  private Location shopperLocation = new Location("reverse");
-
-  public SelectShopperListItemView(Context context) {
-    super(context);
-    init();
-  }
-
-  public SelectShopperListItemView(Context context, AttributeSet attrs, int defStyleAttr) {
-    super(context, attrs, defStyleAttr);
-    init();
-  }
-
-  public SelectShopperListItemView(Context context, AttributeSet attrs) {
-    super(context, attrs);
-    init();
-  }
-
-  public static SelectShopperListItemView inflate(ViewGroup parent) {
-    SelectShopperListItemView itemView = (SelectShopperListItemView) LayoutInflater.from(parent.getContext())
-        .inflate(R.layout.select_shopper_list_item, parent, false);
-    return itemView;
-  }
-
-  private void init() {
-    LayoutInflater.from(getContext()).inflate(R.layout.select_shopper_list_item_children, this, true);
-    ButterKnife.inject(this);
-  }
-
-  public void setShopper(Contact shopper) {
-    shopperLocation.setLatitude(shopper.getLatitude());
-    shopperLocation.setLongitude(shopper.getLongitude());
-    Picasso.with(getContext()).load(shopper.getAvatarUrl()).into(photoView);
-    nameView.setText(shopper.getName());
-    distanceView.setText(
-        String.format(getContext().getString(R.string.select_shopper_distance_format), 0f));
-  }
-
-  public void setUserLocation(Location userLocation) {
-    if (shopperLocation != null) {
-      float distanceInMeters = userLocation.distanceTo(shopperLocation);
-      distanceView.setText(String.format(getContext().getString(R.string.select_shopper_distance_format), distanceInMeters / 1000.0f));
-    }
-  }
-}
