--- app/src/main/res/layout/activity_select_shopper.xml
@@ -1,19 +1,0 @@
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-  <fragment
-      android:id="@+id/select_shopper_map"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:name="com.google.android.gms.maps.MapFragment"
-      />
-  <Button
-      android:id="@+id/select_shopper_go_yourself_button"
-      android:layout_width="match_parent"
-      android:layout_height="48dp"
-      android:text="@string/select_shopper_go_yourself_button"
-      android:layout_alignParentBottom="true"
-      />
-</RelativeLayout>
