--- app/src/main/res/layout/select_shopper_list_item.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.iic.shopingo.ui.request_flow.views.SelectShopperListItemView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    >
-</com.iic.shopingo.ui.request_flow.views.SelectShopperListItemView>
