--- app/src/main/res/layout/select_shopper_list_item_children.xml
@@ -1,44 +1,0 @@
-<merge
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-  <android.support.v7.widget.CardView
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      style="@style/Card"
-      >
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:padding="@dimen/default_spacing"
-        >
-      <ImageView
-          android:id="@+id/select_shopper_list_item_photo"
-          android:layout_marginRight="@dimen/default_spacing"
-          style="@style/Avatar"
-          />
-      <LinearLayout
-          android:orientation="vertical"
-          android:layout_width="0dp"
-          android:layout_height="match_parent"
-          android:layout_weight="1"
-          android:gravity="center_vertical"
-          >
-        <TextView
-            android:id="@+id/select_shopper_list_item_name"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            style="@style/TitleText"
-            />
-        <TextView
-            android:id="@+id/select_shopper_list_item_distance"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            style="@style/BodyText"
-            />
-      </LinearLayout>
-    </LinearLayout>
-  </android.support.v7.widget.CardView>
-</merge>
